Moving composer components to styled components - TooltipList
